usb: dwc3: Fix incorrect PCM value for mult less than 2
For isochronous endpoints with mult less than 2, the
PCM value of the trb->size is set incorrectly.
For case of mult = 1, this is set to 0/-1 and for
mult = 0, this is set to -2.
This is because the initial mult is set to
ep->mult - 1. Fix this by setting the value of
initial value of mult to 2.

Change-Id: Ifa63d8c65564108303ce7c0d6e0cafeccf2ba47d
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>